Move litellm configuration from clients/__init__ into _litellm
litellm globals (telemetry, cache, logging) are now set by
_configure_litellm(), called once when _resolve_backend first selects
the litellm backend. The eager `import litellm` is removed from
clients/__init__.py.

enable_litellm_logging / disable_litellm_logging remain public API
via thin delegates in clients/__init__.py.
